Crate fmt_extra

Structs

AsciiStr
A wrapper around anything that dereferences to a byte slice ([u8]) that displays it's contents as a string using rust-like (& c-like) escapes for non ascii characters

Hs
A wrapper around anything that dereferences to a byte slice ([u8]) that displays it's contents as a hexidicimal string.

Seperated
Display a given iterator over display-ables using by printing each display-able 1 display-able seperator.

SingleQuotedStr
Given a str, display it as a posix-shell style single quoted string with no ambiguity.